+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - Display which theme is the default one in colored output, see #2838 (@sblondon)
 - Add aarch64-apple-darwin ("Apple Silicon") binary tarballs to releases, see #2967 (@someposer)
 - Update the Lisp syntax, see #2970 (@ccqpein)
+- Use bat's ANSI iterator during tab expansion, see #2998 (@eth-p)
 
 ## Syntaxes